# openapi_proposal.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SpotiFLAC Headless API",
    "description": "Application state and download endpoints",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get Global Settings",
        "operationId": "get_global_settings_api_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettings"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Update Global Settings",
        "operationId": "update_global_settings_api_settings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings/profiles/{name}": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get Single Profile",
        "operationId": "get_single_profile_api_settings_profiles__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Create Or Update Profile",
        "operationId": "create_or_update_profile_api_settings_profiles__name__post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppSettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Settings"
        ],
        "summary": "Remove Profile",
        "operationId": "remove_profile_api_settings_profiles__name__delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppSettings": {
        "properties": {
          "output_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Dir"
          },
          "quality": {
            "$ref": "#/components/schemas/AudioQuality",
            "default": "LOSSLESS"
          },
          "services": {
            "items": {
              "$ref": "#/components/schemas/ProviderService"
            },
            "type": "array",
            "title": "Services",
            "default": [
              "tidal",
              "qobuz",
              "deezer"
            ]
          },
          "lyrics_providers": {
            "items": {
              "$ref": "#/components/schemas/LyricsProvider"
            },
            "type": "array",
            "title": "Lyrics Providers",
            "default": [
              "spotify",
              "apple",
              "musixmatch",
              "lrclib",
              "amazon"
            ]
          },
          "filename_format": {
            "type": "string",
            "title": "Filename Format",
            "default": "{title} - {artist}"
          },
          "use_track_numbers": {
            "type": "boolean",
            "title": "Use Track Numbers",
            "default": false
          },
          "use_album_track_numbers": {
            "type": "boolean",
            "title": "Use Album Track Numbers",
            "default": false
          },
          "use_artist_subfolders": {
            "type": "boolean",
            "title": "Use Artist Subfolders",
            "default": false
          },
          "use_album_subfolders": {
            "type": "boolean",
            "title": "Use Album Subfolders",
            "default": false
          },
          "embed_lyrics": {
            "type": "boolean",
            "title": "Embed Lyrics",
            "default": true
          },
          "enrich_metadata": {
            "type": "boolean",
            "title": "Enrich Metadata",
            "default": true
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "AppSettings"
      },
...
      "AudioQuality": {
        "type": "string",
        "enum": [
          "LOSSLESS",
          "HIGH",
          "NORMAL"
        ],
        "title": "AudioQuality"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LyricsProvider": {
        "type": "string",
        "enum": [
          "spotify",
          "apple",
          "musixmatch",
          "lrclib",
          "amazon"
        ],
        "title": "LyricsProvider"
      },
...
      "ProviderService": {
        "type": "string",
        "enum": [
          "tidal",
          "qobuz",
          "deezer",
          "apple",
          "soundcloud",
          "spoti"
        ],
        "title": "ProviderService"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}